% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner_boys.R
\docType{data}
\name{corner_boys}
\alias{corner_boys}
\title{Corner Boys}
\format{
One 13 X 13 matrix (The Nortons), and another one of 8 X 8
(Angelo's boys).
\describe{
    \item{thenortons}{A 13 X 13 matrix of the corner boys in the nortons
    gans. The ones who send the tie are in row, and the ones who received
    the ties are in columns. It represent the line of influence between them.}
    \item{statusnortons}{The status of each corener boys. "1" is higher, and
     "7" is the lowest.}
    \item{angelo}{A 8 X 8 matrix of corner boys included in the Angelo's
    boys. The ones who send the tie are in row, and the ones who received
    the ties are in columns. It represent the line of influence between them.}
    \item{statusangelo}{The status of each corener boys. "1" is higher, and
     "5" is the lowest.}
    \item{freqangelo}{The fraquency of the corner boys, where "1" is
    "frequent", and "2" is for those "infrequently present".}
    }
}
\source{
Whyte, Willian Foote (1993[1943]). Street corner society. The social
structure of an italian slum. Chicago, United States: The University
of Chicago Press.
}
\usage{
data(corner_boys)
}
\description{
This data was as a result of a etnography done in an italian american slum of
Corneville, in Boston, USA. Data collected around 1937-40.
}
\keyword{datasets}
