% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labor_mapuche.R
\docType{data}
\name{labor_mapuche}
\alias{labor_mapuche}
\title{Medieria in Coipuco by Mapuches 1968-69.}
\format{
A list of two 28 X 15 matrices of the people medieria in Coipuco
1968-19.
\describe{
    \item{label_labor}{1: Agriculture in Coipuco. 2: Agriculture with Mapuche.
    3: Agriculture with Chilean. 4: Animal husbandry in Coipuco. 5: Animal
    husbandry with Mapuche. 6: Animal husbandry with Chilean. 7: Poultry and
    pigs in Coipuco. 8: Poultry and pigs with Mapuche. 9: Poultry and pigs
    with Chilean. 10: Firewood and charcoal in Coipuco. 11: Firewood and
    charcoal with Mapuche. 12: Firewood and charcoal with Chilean. 13:
    Occasional in Coipuco. 14: Occasional with Mapuche. 15: Occasional with
    Chilean.}
    \item{labor_owener}{The ties is being owener of the company. The number
    represent how many companies are in total.}
    \item{labor_wp}{The ties is being working partner of the company.
    The number represent how many companies are in total.}
}
}
\source{
Stuchlik, Milan (1976). The life on half share. Mechanisms of
Social Recruitment among the Mapuche of Southern Chile. London: C. Hurst &
Company.
}
\usage{
data(labor_mapuche)
}
\description{
This matrices are part of a study that was conducted at Coipuco, in Chile,
Southamerica, between 1968 and 1970. The data was collected by Milan Stuchlik.
It has been recognized as one of the first studies in Chile and latinamerica
 using social network analysis (Ortiz and Espinosa-Rada, forthcoming).
The complete work was published as a book by Stuchlik (1976), which
described in more detailed the complete etnography.
}
\references{
Ortiz, Francisca and Espinosa-Rada, Alejandro (forthcoming).
Ciencias de redes sociales en America Latina: enfasis y nuevas directrices
en el campo de las ciencias sociales. In Ortiz, Francisca and Espinosa-Rada,
Alejandro (eds.) El analisis de redes sociales en Latinamerica. Madrid,
Spain: Editorial CIS.
}
\keyword{datasets}
