% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations.R
\name{matrix_to_edgelist}
\alias{matrix_to_edgelist}
\title{Transform symmetric matrix to an edge-list}
\usage{
matrix_to_edgelist(A, digraph = FALSE, valued = FALSE, loops = FALSE)
}
\arguments{
\item{A}{A symmetric matrix}

\item{digraph}{Whether the matrix is directed or not}

\item{valued}{Add a third columns with the valued of the relationship}

\item{loops}{Whether the loops are retained or not}
}
\value{
This function return the counts of the dyadic census.
}
\description{
Transform symmetric matrix to an edge-list
}
\examples{

data("bank_room")
matrix_to_edgelist(bank_room$friendship, digraph=TRUE)

}
\author{
Alejandro Espinosa-Rada
}
