% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southern_women.R
\docType{data}
\name{southern_women}
\alias{southern_women}
\title{Southern Women}
\format{
A 6 X 9 incident matrix of women in events.
}
\source{
Davis, A., Gardner, B. B. and M. R. Gardner (1941). Deep South. The University of Chicago Press, Chicago.
}
\usage{
data(southern_women)
}
\description{
This data was collected in 1930 for ethnographic studies by Davis, Gardner and Gardner. It consists of two incident matrices. The first matrix is a group of 18 women attending 14 events. The second matrix is a different group of 6 women attending 9 events.
}
\keyword{datasets}
